+++ FSarchive.php
@@ -1,0 +1,32 @@
+<?php
+/****************** DO NOT REMOVE OR ALTER THIS HEADER ******************************
+*                                                                                   *
+* Product: FSnode                                                                   *
+*    FSnode is a node-based Uniform FileSystem handler, written in PHP. It allows   *
+*    you to access all kind of file systems in the exact same manner, with the same *
+*    simple commands. You can write your web application once and let users switch  *
+*    file system/platform; mount through URI.                                       *
+*                                                                                   *
+* Latest version to download:                                                       *
+*    https://github.com/sentfanwyaerda/FSnode                                       *
+*                                                                                   *
+* Documentation:                                                                    *
+*    http://sent.wyaerda.org/FSnode/                                                *
+*    https://github.com/sentfanwyaerda/FSnode/blob/master/manual/Introduction.md    *
+*                                                                                   *
+* Authors:                                                                          *
+*    Sent fan Wy&aelig;rda (fsnode@sent.wyaerda.org) [creator, main]                *
+*                                                                                   *
+* License: cc-by-nd                                                                 *
+*    Creative Commons, Attribution-No Derivative Works 3.0 Unported                 *
+*    http://creativecommons.org/licenses/by-nd/3.0/                                 *
+*    http://creativecommons.org/licenses/by-nd/3.0/legalcode                        *
+*                                                                                   *
+****************** CHANGES IN THE CODE ARE AT OWN RISK *****************************/
+
+require_once(dirname(__FILE__).DIRECTORY_SEPARATOR.'FSfile.php');
+
+class FSarchive extends FSfile {
+
+}
+?>